requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=1.7.6
shap>=0.44.0
plotly>=5.18.0
streamlit>=1.29.0
fpdf2>=2.7.8
joblib>=1.3.0
pyyaml>=6.0
tqdm>=4.66.0
pytest>=7.4.0

